Our first Android Studio project: Download and Install an APK under program guidance
Of course a normal app can't get permission to install apps, so we just ask Android
to open the downloaded APK via an intent.

Will not work if your SD card is hidden because you have USB set to Memory Card mode.

Will try to make the format dual AndroidStudio-Eclipse later: it can be done!